refactor: extract mastodon-specific code into own file
Preparation for compatibility with e.g. Calckey
